+++ fantasy-oauth2/src/main/java/com/mzlalal/oauth2/service/impl/UserServiceImpl.java
@@ -31,11 +31,11 @@ public class UserServiceImpl extends ServiceImpl<UserDao, UserEntity> implements
         // 查询参数
         QueryWrapper<UserEntity> wrapper = new QueryWrapper<>(po.getEntity());
         // 异步查询总行数 selectList一定要在future之后
-        Future<Integer> future = ThreadUtil.execAsync(() -> baseMapper.selectCount(wrapper));
+        Future<Long> future = ThreadUtil.execAsync(() -> baseMapper.selectCount(wrapper));
         // 查询结果集
         List<UserEntity> entityList = baseMapper.selectList(wrapper);
         // 获取总行数结果
-        Integer count = this.getTotalResult(future, log);
+        Long count = this.getTotalResult(future, log);
         // 返回结果
         return new Page<>(entityList, count, po.getPageInfo());
     }
